audio: ensure effect chain with specific session id is unique
It's possible that tracks with the same session id running on various
playback outputs, which causes effect chain being created on the same
session twice. As a result, the same effect engine will be released
twice as the same context is reused.

Output that has effect chain with same session id is more preferable.

CRs-Fixed: 2232451
Change-Id: I690ea3cb942d1fdc96b46048e271557d48000f43